Tests tidy
